Update scripts to resolve cibuild errors.
The cibuild, lint, manage, test, and update scripts
were updated from recent project that has more
updated STRTA scripts, which resolves a few errors
both at console in the dev environment, and in CI.

A few lines in the test and lint scripts needed to
be commented out to resolve errors caused by the
lack of a frontend. I’ve added TODO comments where
this was done, to assist reenabling these tests later.

docker-compose.ci.yml was causing an error due to an
invalid indentation level. The error below was
resolved by indenting the service stanzas one level
in from the `services` keyword:

You might be seeing this error because you're using
the wrong Compose file version. Either specify a
supported version (e.g "2.2" or "3.3") and place
your service definitions under the `services` key,
or omit the `version` key and place your service
definitions at the root of the file to use version
1.

This also resolves errors flagged by the linting
process. The commented out imports will need to
be uncommented or deleted when the site is developed.