Fix react-remove-scroll-bar warning
This happens when there are multiple modals and each tries to
block scrolling of the parent container, but it has already
been blocked by the previous modal.

This makes it so that only the first one does so, preventing the
warning from appearing in the console.

For reference see: https://github.com/chakra-ui/chakra-ui/issues/6213#issuecomment-1216003840