Add usePreventMapDoubleClick hook
Leaflet seems to be adding capturing event handlers for map events.
This means that stopping propagation from child components of the
map won't prevent the map events from firing. This hook adds event
listeners to the document body above the map with capture set to true
so the events can be intercepted before they get to the map.