Save layer corners on layer remove
While the add event might make more sense, the corners are not loaded at
that time. The reason for the conditional layer._corners check is that
corners are not always loaded when the layer remove event is fired.
For example, the remove event is fired twice while adding the reference
image.