Add State Model & state FK field to User & Utility
Since Utility and State have a many-to-one relationship, with state being a non-nullable FK field on Utility, a default State needed to exist before the creation of the state FK field on Utility. A data migration adding in the NC State row occurs after the State model migration and before the Utility update migration resolves this.