Wire up subissions list page to API
This fetches the boundaries for the current user and shows them
in the table. Each row is clickable to see submission details.
The data is sorted most recent first. Appropriate details are
filled in to the table. There is a spinner shown while fetching.
An error message is displayed in case of failure.