Add useBoundaryId hook
There are two reasons to do this:

The first is that it's less ambiguous than using
`const { id } = useParams()`. This is especially important in deeply
nested components where there are other elements with their own IDs.

Second, if this logic were to change at a later point (perhaps it is
moved into a Context), then it can be updated in one place.