Fix map zooming bug
The primary useEditingPolygon useEffect is called on each map zoom
change because it needs to realign the editing markers on the
polygon. Setting the map to fit bounds within that useEffect caused
the map to be locked to the shape, and did not allow zooming around.

By separating that into a second useEffect, which is called only
once per rendered page, we ensure that the map initializes to be
centered around the shape, but then allows the user to navigate
around.