Show appropriate back button in NavBar
If the user canWrite or canReview, we show a "Save and back"
button. Otherwise, we show a "Back" button.

This is more complex than it needs to be because the NavBar
does not have easy access to the boundary. This is pieceed
together but may not more hardening in the future. In the
case of failure, it should fallback to just saying "Back".