Augment back-end to save reference image files
Add a file field to the reference image model, serializer,
views. Saves the file to S3, under a folder structure
grouping it under the boundary.

Splits the serializer into two: one for creation that has
the file details, another for updating which does not.

Also makes the filename a required field.