Support ingesting zipped shapefile, geojson
If the payload's shape is a zipped shapefile or a geojson,
they will be parsed as shapes and added to the database. In
the case of geojson, the file is parsed and ingested directly.
In the case of zipped shapefiles, we save them to a temporary
directory, then load them via fiona and extract the shape from
within. fiona cannot read files from memory, and the uploaded
files could potentially be really heavy, so doing so helps
manage the load.

These are the only two formats we are supporting at the
moment, more may be added in the future.

This assumes a single Polygon input. Support for MultiPolygons
will be added in #127.