Handle loading shapes before reference images
In some cases, shapes can render before reference images, and
when we try to fit the map bounds to the shape, since the
reference images are not yet loaded there is a crash.

This try block ensures that the crash does not propagate, and
keeps retrying until the map fits once, and then no more.