Have all users go to /submissions by default
The prime case where we don't want to take the user to the
submissions list is when they are a new contributor, adding
a boundary for a utility that doesn't have one already. In
all other cases, we want the user to see the submissions
list.

To make that determination, we need to first fetch the list
of boundaries for the contributor's selected utility, which
happens in the submissions list page. Thus, we redirect all
users there, and make the decision whether to redirect to
/welcome there.