Redirect new users to /welcome from /submissions
Previously, we would redirect all contributors to the /welcome
page. This led to them being welcomed, even if they had already
started a boundary, or submitted one, both cases in which the
/submissions list is a better starting view for them.

The only case when /welcome makes sense is when there are no
boundaries for their selected utility.

Here, we fetch the boundaries for their selected utility, and
iff they are coming from the /login page, we redirect them to
/welcome. Otherwise, they always see the submissions list.