Select related utility/state instead of submission
The utility and state are used later by either this function or the
send email function used later.

Prefetching submissions fetches all submissions, but we need only the
latest here. Since prefetch involves multiple queries anyway, we might
as well let the `latest_submission` cached property do it because it
will limit it to only one.