Update NavBar behavior in Detail page
Previously the Detail NavBar had the Utility Selector and
a Log Out button, which didn't make sense for a detail view.
Now they are replaced with appropriate read only utility
view and a back button that takes the user to the Submissions
List, which is a much more natural placement.

Furthermore, this allows us to differentiate between the List
and Detail NavBars, which will be useful when we add the Export
button.