Add back-end for exporting approved boundaries
This generates a GeoJSON in Postgres and returns a response
that is downloaded as a file.

It is important to do this within Postgres for performance
reasons.

Unfortunately, we do miss out on Django-specific capabilities,
such as getting the "status" of a boundary. Currently we are
only exporting Approved boundaries, but if we were to add
support for exporting All boundaries, we would have to recreate
the status generation logic in SQL.