Fix geojson uploads
Previously we would naively parse any uploaded GeoJSON files,
assuming them all to be polygon geometries. This hardens that,
so that if a user uploads a FeatureCollection or a Feature,
we extract the first Polygon geometry from within that to parse.
If no parseable geometry is found, an error is returned.

Fixes #222