Add management command to bulk import contributors
This management command reads a CSV file from S3, and creates
a contributor user associated with the utilities having given
PWSIDs for each row in the CSV.

This command can be run locally or on AWS using ecsmanage:

    ./scripts/manage create_contributors_from_s3 iow-development-data-us-east-1 csv/test-contributors-success.csv

    ./scripts/manage ecsmanage create_contributors_from_s3 iow-staging-data-us-east-1 csv/test-contributors-success.csv

The output is logged to the console, and also to a file
timestamped to each run. In the case of AWS environments,
that file is also sent to the corresponding logs bucket.

The ECS tasks are given additional permissions to be able to
write to that bucket.

The local log files are added to .gitignore.